Refactor query helpers to traits
